DEVOPSPN-1079

+++ vagrant/test.bash
@@ -1,13 +1,16 @@
 #!/bin/bash
 
 TESTS=(
-"ubuntu12.04"
 "ubuntu14.04"
 "ubuntu16.04"
+"ubuntu18.04"
+"centos7.7"
 )
 OUT="results.txt"
 
-rm $OUT
+if [ -f $OUT ] ; then
+  rm $OUT
+fi
 
 function cleanup {
     vagrant destroy -f
